refactor: Move main.c to src directory
- Delete all code that is now in modules
